Refactored code into folders. Added residual files to .gitignore
